+++ packages/jokul/src/components/logo/index.ts
@@ -1,0 +1,12 @@
+export type { LogoProps } from "./Logo";
+export { Logo } from "./Logo";
+export type { LogoStampProps } from "./LogoStamp";
+export { LogoStamp } from "./LogoStamp";
+export {
+    ForsikringLevertAvFremtind,
+    FraSB1ogDNB,
+    InnovasjonFraFremtind,
+    TeknologiFraFremtind,
+    VartEgetForsikringsselskap,
+    VartForsikringsselskap,
+} from "./text-paths";