+++ packages/jokul/src/components/logo/text-paths/index.ts
@@ -1,0 +1,6 @@
+export { ForsikringLevertAvFremtind } from "./ForsikringLevertAvFremtind";
+export { FraSB1ogDNB } from "./FraSB1ogDNB";
+export { InnovasjonFraFremtind } from "./InnovasjonFraFremtind";
+export { TeknologiFraFremtind } from "./TeknologiFraFremtind";
+export { VartEgetForsikringsselskap } from "./VartEgetForsikringsselskap";
+export { VartForsikringsselskap } from "./VartForsikringsselskap";